Split ValidatorFileRowScopeTest into 2 test files, and changed the name of ValidatorTraitImporterTest. This means each of the data-row validators have their own test file
