+++ trpcultivate_phenotypes/tests/src/Kernel/Validators/ValidatorValueInListTest.php
@@ -4,15 +4,14 @@ namespace Drupal\Tests\trpcultivate_phenotypes\Kernel\Validators;
 use Drupal\tripal_chado\Database\ChadoConnection;
 use Drupal\Tests\tripal_chado\Kernel\ChadoTestKernelBase;
 
- /**
-  * Tests Tripal Cultivate Phenotypes Validator Plugins that apply to a single row
-  * of the input file - the "FILE ROW" scope
-  *
-  * @group trpcultivate_phenotypes
-  * @group validators
-  * @group file_row_scope
-  */
-class ValidatorFileRowScopeTest extends ChadoTestKernelBase {
+/**
+ * Tests the Value In List validator
+ *
+ * @group trpcultivate_phenotypes
+ * @group validators
+ * @group row_validators
+ */
+class ValidatorValueInListTest extends ChadoTestKernelBase {
   /**
    * Plugin Manager service.
    */
@@ -133,49 +132,5 @@ class ValidatorFileRowScopeTest extends ChadoTestKernelBase {
     $validation_status = $instance->validateRow($file_row);
     $this->assertEquals($expected_status, $validation_status['status'], "Value in list validation was expected to fail when provided 1 cell with the same text but wrong case as the one valid value provided.");
     $this->assertStringEndsWith('with >=1 case insensitive match', $validation_status['title'], "Value in list validation title did not specify that a case insensitive match was found.");
-  }
-
-  /**
-   * Test Empty Cell Plugin Validator.
-   */
-  public function testValidatorEmptyCell() {
-
-    // Create a plugin instance for this validator
-    $validator_id = 'empty_cell';
-    $instance = $this->plugin_manager->createInstance($validator_id);
-
-    // Simulates a row within the Trait Importer
-    // Includes 3 types of empty cells
-    $file_row = [
-      'My trait',
-      '',  // No whitespace
-      'My method',
-      ' ', // Single whitespace
-      'My unit',
-      '  ' // Double whitespace
-    ];
-
-    // Case #1: Provide a list of indices for cells that are not empty
-    $expected_status = 'pass';
-    $indices = [ 0, 2, 4 ];
-    $instance->setIndices($indices);
-    $validation_status = $instance->validateRow($file_row);
-    $this->assertEquals($expected_status, $validation_status['status'], "Empty cell validation was expected to pass when provided only non-empty cells to check.");
-
-    // Case #2: Provide a list of indices that includes only empty cells
-    $expected_status = 'fail';
-    $indices = [ 1, 3, 5 ];
-    $instance->setIndices($indices);
-    $validation_status = $instance->validateRow($file_row);
-    $this->assertEquals($expected_status, $validation_status['status'], "Empty cell validation was expected to fail when provided 3 empty cells to check.");
-    $this->assertStringEndsWith(': 1, 3, 5', $validation_status['details'], "Empty cell validation details did not contain the index of the empty cell.");
-
-    // Case #3: Provide a list of indices for the entire row (mixture of 3 empty and 3 non-empty cells)
-    $expected_status = 'fail';
-    $indices = [ 0, 1, 2, 3, 4, 5 ];
-    $instance->setIndices($indices);
-    $validation_status = $instance->validateRow($file_row);
-    $this->assertEquals($expected_status, $validation_status['status'], "Empty cell validation was expected to fail when provided a mixture of 3 empty and 3 non-empty cells to check.");
-    $this->assertStringEndsWith(': 1, 3, 5', $validation_status['details'], "Empty cell validation details did not contain the indices of the empty cells.");
   }
 }
